Customized product page for Fabrics
Added a way for user to buy Fabrics in increments of 0.25 or 0.5. The increment will be set by the merchant and the price of each increment.
If a custom wants to buy 2.5 mtr of Fabric, they can do that. In the backend, 5 quantity of the product will be added to cart.